+++ client/src/components/appbar/index.js
@@ -1,52 +1,20 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { getApiURL } from '../../utils.js';
-import { UserContext } from '../UserContext.js';
+import { getLoginURL } from '../../utils.js';
+import useAuth from '../UserContext.js';
 import { AppBar as MAppBar, Avatar, Box, Button, IconButton, Menu, MenuItem, Toolbar, Tooltip, Typography } from '@mui/material';
 import { Menu as MenuIcon } from '@mui/icons-material';
 import './styles.css';
 
 const AppBar = () => {
-  const { currentUser, _, removeCurrentUser } = useContext(UserContext);
+  const { user, removeCurrentUser } = useAuth();
   const navigate = useNavigate();
   // User menu
   const [anchorEl, setAnchorEl] = React.useState(null);
   const open = Boolean(anchorEl);
 
-  const setupLoginURL = async () => {
-    let url =
-      'https://greatnotes-security-levelup.auth.eu-west-1.amazoncognito.com/login?response_type=code&';
-    url +=
-      'redirect_uri=' +
-      (process.env.NODE_ENV === 'development'
-        ? 'http://localhost:3000'
-        : 'https://great-notes.projects.bbdgrad.com') +
-      '/login';
-    let getClientIdUrl = getApiURL() + '/api/auth/client_id';
-    console.log('getClientIdUrl', getClientIdUrl);
-    url = await fetch(getClientIdUrl)
-      .then(async (res) => {
-        if (!res.ok) {
-          res.json().then((data) => {
-            alert(data);
-          });
-          return;
-        } else {
-          return await res.json().then(async (data) => {
-            url += '&client_id=' + data;
-            return url;
-          });
-        }
-      })
-      .catch((err) => {
-        alert('API down!', err);
-        return;
-      });
-    return url;
-  };
-
   const onLogin = async () => {
-    let url = await setupLoginURL();
+    let url = await getLoginURL();
     if (url === undefined) {
       return;
     }
@@ -75,8 +43,8 @@ const AppBar = () => {
             <MenuIcon />
           </IconButton>
           <nav className="app-bar__nav">
-          {!currentUser && <Button onClick={onLogin}>Login</Button>}
-          {currentUser &&
+          {!user && <Button onClick={onLogin}>Login</Button>}
+          {user &&
             <ul>
               <li>
                 <Link to="/">
@@ -98,8 +66,8 @@ const AppBar = () => {
                     aria-expanded={open ? 'true' : undefined}
                     onClick={handleProfileClick}
                   >
-                    {currentUser.name}
-                    <Avatar src={currentUser.picture} />
+                    {user.name}
+                    <Avatar src={user.picture} />
                   </Button>
                 </Tooltip>
                 <Menu
